removed commented code
