+++ platform/src/test/kotlin/org/stellar/anchor/platform/data/JdbcSep31TransactionTest.kt
@@ -31,29 +31,4 @@ class JdbcSep31TransactionTest {
             }
         """
       .trimIndent()
-
-  //  @Test
-  //  fun `test JdbcSep31Transaction refunds Json conversion`() {
-  //    val txn = JdbcSep31Transaction()
-  //    txn.refundsJson = refundsJsonNoRefundPayment
-  //    // strict is set to false because refundPayments is omitted in txn.refundsJson when it is
-  // set to
-  //    // null.
-  //    JSONAssert.assertEquals(txn.refundsJson, refundsJsonNoRefundPayment, false)
-  //    assertEquals("10", txn.refunds.amountRefunded)
-  //    assertEquals("5", txn.refunds.amountFee)
-  //    assertNull(txn.refunds.refundPayments)
-  //
-  //    txn.refundsJson = refundsJsonWithRefundPayment
-  //    JSONAssert.assertEquals(txn.refundsJson, refundsJsonWithRefundPayment, true)
-  //    assertEquals("10", txn.refunds.amountRefunded)
-  //    assertEquals("5", txn.refunds.amountFee)
-  //    assertEquals(2, txn.refunds.refundPayments.size)
-  //    assertEquals("1", txn.refunds.refundPayments[0].id)
-  //    assertEquals("5", txn.refunds.refundPayments[0].amount)
-  //    assertEquals("1", txn.refunds.refundPayments[0].fee)
-  //    assertEquals("2", txn.refunds.refundPayments[1].id)
-  //    assertEquals("5", txn.refunds.refundPayments[1].amount)
-  //    assertEquals("4", txn.refunds.refundPayments[1].fee)
-  //  }
 }
